Add new script to run chromium without websecurity
Create new user directory in $HOME/.cache/chromium/debug and use that as
the user-data-dir and also set --disable-web-security for debugging
purposed only. (No CORS).
Turned out to be quite effective for light scrolling as no bloat of
configured chromium (extensions, sync, etc) Could add a hotkey to run it
in KDE.