Add plasmabar, barsleep, barwake
plasmabar - calls 'bar' and stores output in ~/.cache/pbar file.
The contents of this file can be printed on the bar using command output
widget. The widget itself can be updated every 1 or 10 milliseconds as
it just cats the output of the file. plasmashell call bar and then calls
barsleep script that sleeps for a fixed amount of time, then repeats
calling bar.

barsleep - sleep of a particular period of time (can argumentise in the
future)

barwake - kills all barsleep instances