Add mc, mc-config to control media using playerctl
mc is a wrapper for playerctl that uses the player passed as $2 or
stored in ~/.config/mc.conf and passes it with -p to playerctl. If both
$2 and the file is absent then doesnt pass -p flag to playerctl and
playerctl by design takes the first player in the list.

mc-config lists all the players available in rofi and then stores
selection in ~/.config/mc.conf file to be used by mc