Update cmdsp and webcam
cmdsp - now shows name of the video devices along with /dev/video. so
that user is not confused. (number allocation changes sometimes).
Due to this cmdsp is totally re-written.

webcam - now calls cmdsp if -q not passed instead of running its own mpv
on a hardcoded /dev/video* file. (Again, as the numbers change).
Also removed the desktop entries from notify-send as its bloat.
Also speficied the absolute path of notification icon of camera-off and
camera-on files